Bug fixes and added printer service in debug mode. Run app in debug mode to access printing services.

+++ app/src/org/grameen/fdp/fragment/LogoutDialogFragment.java
@@ -5,16 +5,22 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.os.Environment;
 import android.preference.PreferenceManager;
 import android.support.v4.app.DialogFragment;
 import android.support.v7.app.AlertDialog;
 import android.util.Log;
+import android.widget.Toast;
 
 import com.salesforce.androidsdk.app.SalesforceSDKManager;
 
 import org.grameen.fdp.R;
 import org.grameen.fdp.activity.BaseActivity;
+import org.grameen.fdp.activity.LandingPageActivity;
 import org.grameen.fdp.activity.LoginActivity;
+import org.grameen.fdp.application.FdpApplication;
+import org.grameen.fdp.utility.Constants;
+import org.grameen.fdp.utility.CustomToast;
 import org.grameen.fdp.utility.DatabaseHelper;
 
 /**
@@ -43,31 +49,37 @@ public class LogoutDialogFragment extends DialogFragment {
                             public void onClick(DialogInterface dialog,
                                                 int which) {
 
-                                prefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
+                                String localPath = Environment.getDataDirectory().getPath() + "/data/" + getActivity().getPackageName() + "/databases/" + Constants.DB_NAME;
+                                String backupPath = FdpApplication.databaseBackupsPath + "/" + Constants.DB_NAME;
 
-                                SalesforceSDKManager.getInstance().logout(getActivity());
+                                if (DatabaseHelper.getInstance(getActivity()).exportDB(localPath, backupPath) != -1) {
 
+                                    Log.i("LOGOUT FRAG", "********  Database backup complete!  **********");
 
-                                Intent intent = new Intent(getActivity(), LoginActivity.class);
-                                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                                intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                                prefs.edit().clear().apply();
-                                try {
-                                    DatabaseHelper.getInstance(getActivity()).deleteAllTables();
-                                    Log.i("LOGOUT FRAG", "DATABASE DELETED");
+                                    prefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
 
-                                    startActivity(intent);
-                                    getActivity().finish();
+                                    SalesforceSDKManager.getInstance().logout(getActivity());
 
-                                } catch (Exception e) {
-                                    e.printStackTrace();
-                                }
+                                    Intent intent = new Intent(getActivity(), LoginActivity.class);
+
+                                    intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
+                                    intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+                                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                                    prefs.edit().clear().apply();
 
 
-/*
+                                    try {
+                                        DatabaseHelper.getInstance(getActivity()).deleteAllTables();
+                                        Log.i("LOGOUT FRAG", "DATABASE DELETED");
 
+                                        startActivity(intent);
+                                        getActivity().finishAffinity();
 
+                                    } catch (Exception e) {
+                                        e.printStackTrace();
+                                    }
+
+                                /*
                                 String accountType = SalesforceSDKManager.getInstance().getClientManager().getAccountType();
 
                                 // Gets a rest client.
@@ -99,7 +111,7 @@ public class LogoutDialogFragment extends DialogFragment {
                                         // Shows everything.
                                     }
                                 });*/
-
+                                }
 
                             }
                         })
